backend/requirements.txt
fastapi>=0.111
pytest>=8.2
pytest-asyncio>=0.23
httpx>=0.27
pydantic-settings>=2.2
python-multipart>=0.0.9
uvicorn[standard]>=0.29
asyncpg>=0.29
pyserial-asyncio>=0.6
mgrs>=1.4
packaging>=24.0
python-dotenv>=1.0
python-jose[cryptography]>=3.3
bcrypt>=4.1
pandas>=2.2
openpyxl>=3.1
weasyprint>=62
